Fixed major bugs with teams and procedures

+++ TournamentOrganizer/Commands/TournamentCommands/DeleteTournamentCommand.cs
@@ -1,4 +1,6 @@
-﻿using TournamentOrganizer.BusinessLayer.Service;
+﻿using System.Windows;
+using TournamentOrganizer.BusinessLayer.Service;
+using TournamentOrganizer.UI.Validation.TabItemTournamentValidation;
 using TournamentOrganizer.UI.VeiwModels;
 
 namespace TournamentOrganizer.UI.Commands.TournamentCommands
@@ -16,6 +18,13 @@ namespace TournamentOrganizer.UI.Commands.TournamentCommands
 
         public override void Execute(object parameter)
         {
+            if (!TabItemTournamentValidation.DeleteTournamentValidation(_viewModel.TournamentParticipants))
+            {
+                MessageBox.Show("Невозможно удалить турнир пока в нем учавствуют люди",
+                           "Ошибка ",
+                           MessageBoxButton.OK);
+                return;
+            }
             _tournamentService.DeleteTournament(_viewModel.SelectedTournament.Id);
             _viewModel.Tournaments.Remove(_viewModel.SelectedTournament);
         }
